Remove EvidenceSubmission directory/namespace
